pyproject.toml
[tool.poetry]
name = "talkie"
version = "0.1.0"
description = "A Python package for conversational AI interactions"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "talkie", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pinecone = "^5.3.1"
openai = "^1.54.3"
chromadb = "^0.5.18"
xxhash = "^3.5.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
jupyterlab = "^4.3.0"
prettyprint = "^0.1.5"
scikit-learn = "^1.5.2"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
talkie = "talkie.cli.main:main"
quick = "talkie.main:quick"
